requirements.txt
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
cssselect==1.2.0
Django==5.0.4
fake-useragent==1.5.1
idna==3.7
importlib_metadata==7.1.0
lxml==5.2.1
lxml_html_clean==0.1.1
meta_ai_api==1.1.7
parse==1.20.1
pyee==11.1.0
pyppeteer==2.0.0
pyquery==2.0.0
requests==2.31.0
requests-html==0.10.0
soupsieve==2.5
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
zipp==3.18.1
feedparser
psycopg2-binary
gunicorn
djangorestframework
pprintpp
objprint
nose